# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Forma única de todos los errores. La UI la traduce a un mensaje por `code`.",
        "properties": {
          "code": {
            "description": "Identificador estable del error, p. ej. NotFoundError",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Datos extra del error",
            "title": "Detail"
          },
          "message": {
            "description": "Mensaje en español, ya legible",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "AuditEntryOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "prevHash": {
            "title": "Prevhash",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "at",
          "actor",
          "action",
          "target",
          "prevHash",
          "hash"
        ],
        "title": "AuditEntryOut",
        "type": "object"
      },
      "AuditPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AuditPage",
        "type": "object"
      },
      "AuditVerificationOut": {
        "properties": {
          "brokenAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brokenat"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "entries",
          "valid"
        ],
        "title": "AuditVerificationOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend del orquestador multiagente. El dominio gira alrededor de un Requerimiento: adjuntos, plan de agentes, ejecuciones y entregables.",
    "title": "Orquestador de revisión de requerimientos",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/audit": {
      "get": {
        "operationId": "index_api_audit_get",
        "parameters": [
          {
            "description": "Requerimiento o ejecución",
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Requerimiento o ejecución",
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Index",
        "tags": [
          "auditoría"
        ]
      }
    },
    "/api/audit/verify": {
      "get": {
        "description": "Recalcula la cadena y dice si alguien tocó una entrada.",
        "operationId": "verify_api_audit_verify_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerificationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Verify",
        "tags": [
          "auditoría"
        ]
      }
    },
...
  }
}